feat(g-base): move util methods of Group to Container for better reuse of Canvas
