+++ docs/api/canvas.zh.md
@@ -1,11 +1,15 @@
 ---
 title: Canvas 画布
-order: 1
+order: 0
+redirect_from:
+  - /zh/docs/api
 ---
 
 ## 属性
 
-### 通用的 [元素属性](/zh/docs/api/element#通用属性)
+### [元素属性](/zh/docs/api/general/element#属性)
+
+### [容器属性](/zh/docs/api/general/container#属性)
 
 ### width: number
 
@@ -39,7 +43,9 @@ export type Renderer = 'canvas' | 'svg';
 
 ## 方法
 
-### 通用的 [元素方法](/zh/docs/api/element#通用方法)
+### [元素方法](/zh/docs/api/general/element#方法)
+
+### [容器方法](/zh/docs/api/general/container#方法)
 
 ### isCanvas()
 
@@ -61,105 +67,9 @@ export type Renderer = 'canvas' | 'svg';
 
 - 设置画布的 cursor 样式，其中 `Cursor` 为样式类型，可参考 [MDN 文档](https://developer.mozilla.org/en-US/docs/Web/CSS/cursor)。
 
-### getShapeBase()
-
-- 获取图形的基类；
-
-### getGroupBase()
-
-- 获取图形分组的基类；
-
-### add(element: Element)
-
-- 添加并返回元素，其中 `element` 为构造好的元素对象；
-
-### addShape(cfg: ShapeCfg)
-
-- 添加并返回图形，其中 `cfg` 为图形配置项，具体说明如下:
-
-```js
-type ShapeCfg = {
-  /* 图形 id */
-  id?: string,
-  /* 图形名称 */
-  name?: string,
-  /* 图形类型 */
-  type?: string,
-  /* 绘图属性 */
-  attrs: {
-    [key: string]
-  },
-  /* 层次索引，决定绘制的先后顺序 */
-  zIndex?: number,
-  /* 是否可见 */
-  visible?: boolean,
-  /* 是否可以拾取 */
-  capture?: boolean,
-  /* 其他属性 */
-  [key: string]: any,
-};
-```
-
-### addShape(type: string, cfg: ShapeCfg)
-
-- 添加并返回图形，其中:
-  - `type`: 图形类型；
-  - `cfg`: 图形配置项，详情同上；
-
-### addGroup()
-
-- 添加并返回一个默认的图形分组；
-
-### addGroup(cfg: GroupCfg)
-
-- 添加并返回图形，其中 `cfg` 为图形分组配置项，具体说明如下：
-
-```js
-type ShapeCfg = {
-  /* 图形分组 id */
-  id?: string,
-  /* 图形分组名称 */
-  name?: string,
-  /* 图形分组的绘图属性 */
-  attrs: {
-    [key: string]
-  },
-  /* 层次索引，决定绘制的先后顺序 */
-  zIndex?: number,
-  /* 是否可见 */
-  visible?: boolean,
-  /* 是否可以拾取 */
-  capture?: boolean,
-  /* 其他属性 */
-  [key: string]: any,
-};
-```
-
-### addGroup(groupClass: Group, cfg: ShapeCfg)
-
-- 添加并返回特定类型的图形分组，其中:
-  - `groupClass`: 图形分组 Class；
-  - `cfg`: 图形分组配置项，详情同上；
-
-### getParent()
-
-- 获取父元素，对于 Canvas 来说为 `null`；
-
-### getChildren()
-
-- 获取所有的子元素；
-
 ### getShape(x: number, y: number);
 
 - 根据画布坐标 `x` 和 `y` 获取对应的图形；
-
-### sort()
-
-- 子元素按照 `zIndex` 属性进行排序；
-
-### clear()
-
-- 清空所有的子元素；
 
 ### changeSize(width: number, height: number)
