+++ docs/api/general/container.en.md
@@ -1,71 +1,19 @@
 ---
-title: Canvas
-order: 0
-redirect_from:
-  - /en/docs/api
+title: Container 容器
+order: 1
 ---
 
 ## 属性
 
 ### [元素属性](/zh/docs/api/general/element#属性)
 
-### [容器属性](/zh/docs/api/general/container#属性)
-
-### width: number
-
-- 画布宽度；
-
-### height: number
-
-- 画布高度；
-
-### container: number | HTMLElement
-
-- 画布容器，可以是容器 `id` 或者 DOM 元素；
-
-### renderer: Renderer (只读属性)
-
-> ⚠️ 注意，该属性为只读属性，不能动态修改。
-
-- 当前使用的渲染引擎，其中 `Renderer` 的类型为:
-
-```ts
-export type Renderer = 'canvas' | 'svg';
-```
-
-### pixelRatio: number
-
-- 画布大小和所占 DOM 宽高的比例，一般可以使用 `window.devicePixelRatio`，通常情况下无需手动设置该属性；
-
-### cursor: Cursor
-
-- 画布的 cursor 样式，其中 `Cursor` 为样式类型，可参考 [MDN 文档](https://developer.mozilla.org/en-US/docs/Web/CSS/cursor)。
-
 ## 方法
 
 ### [元素方法](/zh/docs/api/general/element#方法)
 
-### [容器方法](/zh/docs/api/general/container#方法)
-
 ### isCanvas()
 
 - 是否为画布；
-
-## getRenderer(): Renderer
-
-- 获取渲染引擎，其中 `Renderer` 的类型为:
-
-```ts
-export type Renderer = 'canvas' | 'svg';
-```
-
-## getCursor(): Cursor
-
-- 获取画布的 cursor 样式，其中 `Cursor` 为样式类型，可参考 [MDN 文档](https://developer.mozilla.org/en-US/docs/Web/CSS/cursor)。
-
-## setCursor(cursor: Cursor)
-
-- 设置画布的 cursor 样式，其中 `Cursor` 为样式类型，可参考 [MDN 文档](https://developer.mozilla.org/en-US/docs/Web/CSS/cursor)。
 
 ### getShapeBase()
 
@@ -149,7 +97,7 @@ type ShapeCfg = {
 
 ### getParent()
 
-- 获取父元素，对于 Canvas 来说为 `null`；
+- 获取父元素；
 
 ### getChildren()
 
@@ -167,18 +115,32 @@ type ShapeCfg = {
 
 - 清空所有的子元素；
 
-### changeSize(width: number, height: number)
+## 工具方法
 
-- 修改画布大小；
+### getFirst()
 
-### getPointByClient(clientX: number, clientY: number)
+- 获取第一个子元素；
 
-- 根据窗口坐标，获取对应的画布坐标，返回类型为 `{ x: number, y: number }`；
+### getLast()
 
-### getClientByPoint(x: number, y: number)
+- 获取最后一个子元素；
 
-- 根据画布坐标，获取对应的窗口坐标，返回类型为 `{ x: number, y: number }`；
+### getCount()
 
-### draw()
+- 获取子元素的数量；
 
-- 绘制方法，在 `自动渲染` 模式下无需手动调用，在 `手动渲染` 模式下需要手动调用；
+### findAll(fn: element => boolean)
+
+- 查找所有匹配的元素，其中 `fn` 为查找函数；
+
+### find(fn: element => boolean)
+
+- 查找第一个符合条件的元素，其中 `fn` 为查找函数；
+
+### findById(id: string)
+
+- 根据元素 `id` 查找元素；
+
+### findAllByName(name: string)
+
+- 根据元素名称查找元素列表；
