+++ docs/api/general/element.en.md
@@ -1,11 +1,11 @@
 ---
-title: Element 元素
+title: Element
 order: 0
 redirect_from:
-  - /zh/docs/api
+  - /en/docs/api/general
 ---
 
-# 通用属性
+# 属性
 
 ## id: string
 
@@ -35,7 +35,7 @@ redirect_from:
 
 - [绘图属性](/zh/docs/api/shape/attrs)；
 
-# 通用方法
+# 方法
 
 ## 属性类方法
