+++ docs/api/group.en.md
@@ -1,15 +1,15 @@
 ---
 title: Group
-order: 2
+order: 1
 ---
 
 ## 属性
 
-### 通用的 [元素属性](/en/docs/api/element#通用属性)
+### [元素属性](/en/docs/api/general/element#属性)
 
 ## 方法
 
-### 通用的 [元素方法](/en/docs/api/element#通用方法)
+### [元素方法](/en/docs/api/general/element#方法)
 
 ### isCanvas()
